requirements.txt
Pillow==10.1.0
Pillow==10.4.0
streamlit==1.37.1
